Move UXP installer to dedicated install package and refactor to manager
Moves the UXP installer to an agnostic install package and does some
minor refactoring to refer to the Installer interface as a Manager. This
is more reflective of the purpose, as the install.Manager is actually
responsible for installing, uninstalling, and managing installations.
